style(theme): renamed files to kebab-case format
