+++ packages/theme/src/tokens/index.ts
@@ -3,12 +3,12 @@ import { blurs } from "./blurs"
 import { borders } from "./borders"
 import { breakpoints } from "./breakpoints"
 import { colors } from "./colors"
-import { fontSizes } from "./fontSizes"
-import { fontWeights } from "./fontWeights"
+import { fontSizes } from "./font-sizes"
+import { fontWeights } from "./font-weights"
 import { fonts } from "./fonts"
 import { gradients } from "./gradients"
-import { letterSpacings } from "./letterSpacings"
-import { lineHeights } from "./lineHeights"
+import { letterSpacings } from "./letter-spacings"
+import { lineHeights } from "./line-heights"
 import { radii } from "./radii"
 import { shadows } from "./shadows"
 import { sizes } from "./sizes"
